% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{addCluster}
\alias{addCluster}
\title{Makes a table with the coordinates of the clusters}
\usage{
addCluster(granges, indexes, prev, cluster, type)
}
\arguments{
\item{dir}{the directory that contains the *hybrids.hyb files}

\item{clustering}{The output from the iGraph function cluster_walktrap for the (made with adjacency matrix input)}

\item{highest_clusters}{The cluster you are interested in keeping}

\item{left}{list created with hybToGRanges (but just the left section of the list)}

\item{right}{list created with hybToGRanges (but just the right section of the list)}
}
\value{
A table of clusters and coordinates
}
\description{
Does the same as printClusters but is a lot faster and does not create plots
of each cluster
}
\examples{
plottingList[[i]][[k]]  = printClustersFast(dir,clustering, highest_clusters, chimeraList[["sample"]][["left"]], chimeraList[["sample"]][["right"]])
}
