% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{compareKnown}
\alias{compareKnown}
\title{compareKnown}
\usage{
compareKnown(trimmedClusters, knownMat, rna, type, ...)
}
\arguments{
\item{trimmedClusters}{a \code{comradesClusteredDataSet} object, 
run \code{trimClusters()} on the  \code{comradesClusteredDataSet} first}

\item{knownMat}{}

\item{rna}{}

\item{type}{}
}
\value{
Returns a \code{comradesClusteredDataSet} object

The 3 attributes matrixList, clusterTableList and clusterGrangesList 
will gain the \code{types} "superClusters" and "trimmedClusters"
}
\description{
This method compares the current object to a know structure.run 
\code{trimClusters()} on the  \code{comradesClusteredDataSet} first
}
