% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesClusteredDataSet.R
\docType{class}
\name{comradesClusteredDataSet-class}
\alias{comradesClusteredDataSet-class}
\alias{comradesClusteredDataSet}
\title{comradesClusteredDataSet}
\usage{
comradesClusteredDataSet(cds, cores)
}
\description{
Desc
}
