% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesFoldedDataSet.R
\docType{class}
\name{comradesFoldedDataSet-class}
\alias{comradesFoldedDataSet-class}
\alias{comradesFoldedDataSet}
\title{comradesFoldedDataSet}
\usage{
comradesFoldedDataSet(cdsObject, rna, rnaRefs)
}
\description{
Desc
}
