% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesOO.R
\docType{package}
\name{comradesOO}
\alias{comradesOO}
\title{comradesOO: A package for analysing a COMNRADES dataset}
\description{
The package consists of 3 main objects:
   comradesDataSet, comradesClusteredDataSet annd comradesFoldedDataSet
}
\section{\code{\link{comradesDataSet}}}{
NA
}

\section{\code{\link{comradesClusteredDataSet}}}{
NA
}

\section{\code{\link{comradesFoldedDataSet}}}{
NA
}

