% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesDataSet.R
\docType{class}
\name{comradesDataSet}
\alias{comradesDataSet}
\title{comradesDataSet}
\usage{
comradesDataSet(rnas, hybDir, sampleTable, group, rnaSize)
}
\arguments{
\item{rnas}{desc}

\item{hybDir}{desc}

\item{group}{desc}

\item{rnaSize}{desc}

\item{sampletable}{desc}
}
\value{
A comradesDataSet object.
}
\description{
\code{comradesDataSet} Is used to store the input meta-data, data and create 
a framework for the storage of results. Whilst creating the object, 
the original hyb files are also filtered for the RNA of interest. 
The user must supply the directory of the input files, the rna of interest 
and a sampleTable.
}
\examples{


}
\references{
See \url{http://www-huber.embl.de/users/anders/HTSeq} for hyb
}
