% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesDataSetMethodsAndHelpers.R
\name{featureInfo}
\alias{featureInfo}
\title{featureInfo}
\usage{
featureInfo(cds, file, ...)
}
\arguments{
\item{cds}{a \code{comradesClusteredDataSet} object}

\item{file}{a folder location to write the table of statistics for RNAS}
}
\description{
this method prints some feature level statistics
}
