% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesFoldedDataSetMethodsAndHelpers.R
\name{findBasePairsRNAfold}
\alias{findBasePairsRNAfold}
\title{findBasePairsRNAfold}
\usage{
findBasePairsRNAfold(startPos, endPos, seqs, fasta)
}
\arguments{
\item{fasta}{\code{rnaRefs}}

\item{startPos1}{Start of the cluster side x}

\item{endPos1}{End of the cluster side x}

\item{seq1}{Sequence of x}
}
\value{
A table of clusters and coordinates with folds
}
\description{
Folds the clusters using Vienna RNA duplex
}
\examples{

}
