% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonHelpersAndMethods.R
\name{getAdjacancyMat}
\alias{getAdjacancyMat}
\title{Makes and adjacency matrix list (for clustering)}
\usage{
getAdjacancyMat(hybGranges, nucletideOrPerc, cutoff)
}
\arguments{
\item{hybGranges}{list created with hybToGRanges (but just the gap section of the list)}

\item{nucletideOrPerc}{measure difference by percentage or nucleotides}

\item{cutoff}{The maximum difference before giving these two gaps 0}
}
\value{
A list of Adjacancy matrices
}
\description{
Makes and adjacency matrix list (for clustering)
}
\examples{
 adjcancyMat =  getAdjacancyMat(hybGranges[["sample"]][["gap"]],"nucleotide", 5);
}
