% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getMatrices}
\alias{getMatrices}
\title{Make a matrix of contact interactions}
\usage{
getMatrices(hybList, rna, size)
}
\arguments{
\item{hybList}{the original hybList created with readHybFiles or subsetHybList}

\item{rna}{the rna of interest that you want to subset}
}
\value{
A list of matrices
}
\description{
Function used to create a list of matrices for plotting with
plotMatrixList or plotMatrixListFull, the output list will be same as the
input except for an extra list layer for the specific RNA
}
\examples{
hybMatList <- subsetHybList(hybList, "myRNA");
}
