% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesDataSetMethodsAndHelpers.R
\name{topTranscripts}
\alias{topTranscripts}
\title{topTranscripts}
\usage{
topTranscripts(cds, ntop = 10, ...)
}
\arguments{
\item{cds}{a \code{comradesClusteredDataSet} object}
}
\value{
Returns a \code{comradesClusteredDataSet} object

The 3 attributes matrixList, clusterTableList and clusterGrangesList 
will gain the \code{types} "superClusters" and "trimmedClusters"
}
\description{
This method prints the top transcripts that have the most duplexes 
assigned
}
