% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{trimClusters}
\alias{trimClusters}
\title{trimClusters}
\usage{
trimClusters(clusteredCds, ...)
}
\arguments{
\item{clusteredCds}{a \code{comradesClusteredDataSet} object}
}
\value{
Returns a \code{comradesClusteredDataSet} object

The 3 attributes matrixList, clusterTableList and clusterGrangesList 
will gain the \code{types} "superClusters" and "trimmedClusters"
}
\description{
This method trims the clusters
}
