% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesFoldedDataSet.R
\docType{class}
\name{comradesFoldedDataSet-class}
\alias{comradesFoldedDataSet-class}
\alias{comradesFoldedDataSet}
\title{comradesFoldedDataSet}
\usage{
comradesFoldedDataSet(cdsObject, rna, rnaRefs)
}
\arguments{
\item{cdsObject}{comradesDataSet object created with comradesDataSet}

\item{rna}{string - a single RNA to analyse}

\item{rnarefs}{named List - a list with named elements that correspond to the 
.RNA of interest. The element of the list must be a fasta file that has 
been read with \code{seqinr::read.fasta()}}
}
\description{
This object is a parent of comradesClusteredDataSet with one extra attribute
clusterTableFolded.
}
\section{Slots}{

\describe{
\item{\code{clusterTableFolded}}{table - a table similar to the \code{clusterTableList}
it contains coordinates of the clusters along with vienna format fold and 
RNA sequences for each cluster}

\item{\code{clusterTableList}}{List - Follows the pattern for list slots of comradesDataSet
objects, \code{matrixList(cds)[[rna]][[type]][[sample]]}. contains a table
with coordinates and information about the clusters identified}

\item{\code{clusterGrangesList}}{List - Follows the pattern for list slots of comradesDataSet
objects, \code{matrixList(cds)[[rna]][[type]][[sample]]}. contains GRanges 
objects of the original suplexes with their cluster membership
#' @slot sampleTable table - Column names; fileName, group (s or c),
 sample (1,2,3, etc), sampleName (must be unique)}

\item{\code{rnas}}{vector - A vector of RNA names to analyse}

\item{\code{group}}{list - This is made from the a sample table during object 
creation, it is a list with two vector elements ("c","s") containing the 
indexes of the sampleTable that have "c" or "s" in the group column.}

\item{\code{matrixList}}{List - Follows the pattern for list slots of comradesDataSet
objects, \code{matrixList(cds)[[rna]][[type]][[sample]]}. Contains a set
of contact matrices, each cell contains the number of duplexes identified 
for position x,y.}

\item{\code{hybFiles}}{List - Follows the pattern for list slots of comradesDataSet
objects, \code{hybFiles(cds)[[rna]][[type]][[sample]]}. Contains a set of 
tables, these are the original Hyb files that were read in.}
}}

