% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesDataSetMethodsAndHelpers.R
\name{featureInfo}
\alias{featureInfo}
\title{featureInfo}
\usage{
featureInfo(cds, ...)
}
\arguments{
\item{cds}{a \code{comradesClusteredDataSet} object}
}
\value{
Returns a \code{comradesClusteredDataSet} object

The 3 attributes matrixList, clusterTableList and clusterGrangesList 
will gain the \code{types} "superClusters" and "trimmedClusters"
}
\description{
this method prints some feature level statistics
}
