% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesDataSetMethodsAndHelpers.R
\name{subsetHybList2}
\alias{subsetHybList2}
\title{Subset a list of hyb files}
\usage{
subsetHybList2(hybList, min, max, length)
}
\arguments{
\item{hybList}{the original hybList created with readHybFiles}

\item{min}{the rna of interest that you want to subset}

\item{max}{The number of randomly subsetted chimeric reads you need}

\item{length}{The number of randomly subsetted chimeric reads you need}
}
\value{
A list of subsetted hyb files
}
\description{
Function used to subset a list of hyb data created by readHybFiles
This function produces the same size list as before but with an
it returns ONLY the rna of interest and also
the number parameter specifies the number of chimeras to return
this is helpful when you want to directly compare libraries with
different sizes.
}
\examples{
hybListSubset <- subsetHybList(hybList, "myRNA" ,  10000);
}
